Deprecate LBANN launcher with LC-specific optimizations
lbann.contrib.lc.launcher.run is deprecated by lbann.contrib.launcher.run.
